Refactor client logic for GitHub interactions
Description:
I opted to refactor out all *github.Client logic from the package
and instead wrap it all in an interface which should help in unit
testing/decoupling the code. Now the ingestor unit tests will have
a single location (server_test.go) where the newClient function will
be overridden.

Additional githubService methods can be implemented as additional
GitHub interactions are required by the production code. These
should all (likely) return various structs/errors and when they are
faked in unit tests should be pretty common across the individual
function unit tests (since they're wrapping similar library logic).

I did opt to use a type assertion in the worker.go code to get the
*github.Client struct - we could swap this for another getter method
on githubService (but I just didn't want to write it :grinning:).

Lastly, I changed all ID-related fields to int64 for uniformity. It
seems to be the style in the go-github library as well.

Note: WIP check-in includes incomplete code
Note: Some unit tests are implemented and appear to be passing

+++ ingestor/collector.go
@@ -6,7 +6,7 @@ import "github.com/google/go-github/github"
 type heuprInstallation struct {
 	ID              *int64       `json:"id,omitempty"`
 	Account         *github.User `json:"account,omitempty"`
-	AppID           *int         `json:"app_id,omitempty"`
+	AppID           *int64       `json:"app_id,omitempty"`
 	AccessTokensURL *string      `json:"access_tokens_url,omitempty"`
 	RepositoriesURL *string      `json:"repositories_url,omitempty"`
 	HTMLURL         *string      `json:"html_url,omitempty"`
